move fft.py and signal.py to paddle/, fix typos
